Allow modifier-only row for tall CSV format
If a modifier is encoded without an item or service, then a description
and one of the following is the minimum information required:
additional_payer_notes, standard_charge | negotiated_dollar,
standard_charge | negotiated_percentage, or standard_charge |
negotiated_algorithm.

When a row is a modifier row, other conditionals do not apply.

The current implementation only supports the tall CSV format.