Provide more specific message for invalid state code in header
When validating the header columns, assume that a column name that
contains one pipe and matches "license_number" before the pipe is
probably the license number column. This means that the part of the name
after the pipe should be a valid state code. If not, show an error
message indicating that the state code is not valid.