Conditional checks 5-8 for CSV formats (#11)
* min and max required when dollar is not empty

If there is a "payer specific negotiated charge" encoded as a dollar
amount, there must be a corresponding valid value encoded for the
deidentified minimum and deidentified maximum negotiated charge data.

* estimated amount is required when dollar is empty

If a "payer specific negotiated charge" can only be expressed as a
percentage or algorithm, then a corresponding "Estimated Allowed Amount"
must also be encoded. Required beginning 1/1/2025.

* drug measurement fields are required when code is NDC

If code type is NDC, then the corresponding drug unit of measure and
drug type of measure data element must be encoded. Required beginning
1/1/2025.

* Allow modifier-only row for tall CSV format

If a modifier is encoded without an item or service, then a description
and one of the following is the minimum information required:
additional_payer_notes, standard_charge | negotiated_dollar,
standard_charge | negotiated_percentage, or standard_charge |
negotiated_algorithm.

When a row is a modifier row, other conditionals do not apply.

The current implementation only supports the tall CSV format.

* Add conditional checks for wide CSV format

The same conditional checks for the tall CSV format also apply to the
wide CSV format. Their implementations are different because the wide
format has payer-specific columns.

* remove commented-out code