Set warning flag in JSON for conditions applicable in 2025
Some rules in the schema are not enforced until 2025. If these rules
create errors, mark them as warnings. The number of warnings and errors
are counted separately. If the maximum number of warnings is reached,
further warnings are discarded, but validation continues. If all
ValidationErrors are warnings, the file is considered valid.