Remove byte-order mark from JSON stream and first CSV chunk (#53)
* Remove BOM from JSON stream

When JSON input is a readable stream, a byte-order mark may be present
at the beginning of the stream. When reading the first chunk of the
stream, check for the presence of the byte-order mark. Remove it if it
is present.

* Remove BOM when parsing CSV

When parsing a CSV from a source that includes a byte-order mark (BOM),
the BOM is present at the time the parser attempts to determine if the
first column name is quoted. When the BOM is present, the parser does
not recognize that the column is quoted, resulting in a failure to match
an expected column name. Remove the BOM (if present) from the first
chunk so that a quoted column name will be recognized and parsed as a
quoted value.

* minor refactor to DRY up BOM


---------