Add validateColumns tests
Fix OBOB for code count. Add tests for incorrect use of code columns.

Start filling out tests for validateRow tall

Adding "normalized columns" so that we can easily find the location of
elements, but keep the original formatting for message purposes.

Define some of the tests for validating a row in tall mode. Fields
tested are description, setting, and the drug information fields.

Add tests for code pairs and non-payer specific charges

add tests for payer specific negotiated charge

Add "other" methodology conditional

Still messing around a bit with types and the structure of the
validation tree.

Add tests for requiring at least one charge

I'm starting to feel uneasy about some of these implementations. Might
try to change up how they go so I have less actual logic in the
predicate functions.

Change up validation types and partial function application

Individual validation functions are redefined so that the arguments are
listed in a useful order for binding. This means that when actually
validating the row, the only values to pass in are the data object and
the row index.

Add tests for dollar amount requiring min and max

Remove reset method, since it's really just for convenience, and it
doesn't actually make things that much more convenient.