Renaming, refactoring, reorganizing.
Many constants are moved around so that there's fewer total definitions
in the CsvValidator file.
Bind the CsvValidator instance to the data callback so that the function
will have access to any other properties it needs.
Fill out todo tests for JsonValidator. Make JsonValidator's schema
properties public, both for testing and because it might be useful to be
able to modify them in an inherited class.
Delete some unused constants and types.