Fix extraneous error from prescription drug requirement
The conditional schema for requiring drug information when an NDC code
is present was written such that it would also be applied if the code
information property was not present. Add required keywords to the
conditional schema so that these superfluous errors do not appear.

Fix representations of Jan 2025 errors in tests.