JSON validation callback enhancements
Add arguments to type definition for JSON data callback so that
information about the path can be provided to the callback.
Add metadata callback to JSON validator. The metadata callback, if
defined, is called once with the metadata object and any errors from
validating that object against the metadata schema.
Add errors to top level export to facilitate writing data callbacks.