Collect alerts (#69)
When validating, collect alerts in addition to errors. The maximum for errors also applies to alerts. If the amount of alerts reaches the maximum, validation will continue. The only alert currently implemented is for the presence of a value of `999999999` for estimated amounts. The validate methods return the list of alerts alongside the list of errors.