Validate 2026 rule (#77)
* Refactor validator classes

The current state of this branch is a work-in-progress: lots of
incomplete implementations and tests. Use this branch to work towards a
complete implementation of the refactored validator classes.

* Add validateColumns tests

Fix OBOB for code count. Add tests for incorrect use of code columns.

Start filling out tests for validateRow tall

Adding "normalized columns" so that we can easily find the location of
elements, but keep the original formatting for message purposes.

Define some of the tests for validating a row in tall mode. Fields
tested are description, setting, and the drug information fields.

Add tests for code pairs and non-payer specific charges

add tests for payer specific negotiated charge

Add "other" methodology conditional

Still messing around a bit with types and the structure of the
validation tree.

Add tests for requiring at least one charge

I'm starting to feel uneasy about some of these implementations. Might
try to change up how they go so I have less actual logic in the
predicate functions.

Change up validation types and partial function application

Individual validation functions are redefined so that the arguments are
listed in a useful order for binding. This means that when actually
validating the row, the only values to pass in are the data object and
the row index.

Add tests for dollar amount requiring min and max

Remove reset method, since it's really just for convenience, and it
doesn't actually make things that much more convenient.

* Adding more CSV 2.2.0 tests

Add tests for modifier-only rows.
Add tests for requiring estimated allowed amount when payer-specific
percentage or algorithm, but not dollar amount, are present.
Add tests for requiring drug information when NDC code type is used.

* adding tests for CSV wide format

Add tests for payer specific negotiated charge conditional.
Add tests for "other" methodology conditional.
Add tests for requiring at least one charge.

* fix imports and jest config to work properly as esmodule

* Additional tests for CSV Wide

Add tests for payer-specific dollar amount requiring a min and max.
Add tests for modifier row additional information.
Add tests for requiring an estimated allowed amount for a percentage or
algorithm.
Add tests for requiring drug information when code type is NDC.

* Add tests for main CSV validate function

Build row validators after rows are determined.
Add list of available versions to InvalidVersionError.

* Add tests for all CSV versions

Tests that depend on a specific version of the requirements are split
into separate files, since the main test file was growing too large.

For pre-v2.2.0 versions, there are no modifier rows. Therefore, code
information is always required.

* Renaming, refactoring, reorganizing.

Many constants are moved around so that there's fewer total definitions
in the CsvValidator file.
Bind the CsvValidator instance to the data callback so that the function
will have access to any other properties it needs.
Fill out todo tests for JsonValidator. Make JsonValidator's schema
properties public, both for testing and because it might be useful to be
able to modify them in an inherited class.
Delete some unused constants and types.

* Add tests for remaining CSV errors

* Provide errors to CSV data callback

* JSON validation callback enhancements

Add arguments to type definition for JSON data callback so that
information about the path can be provided to the callback.
Add metadata callback to JSON validator. The metadata callback, if
defined, is called once with the metadata object and any errors from
validating that object against the metadata schema.
Add errors to top level export to facilitate writing data callbacks.

* add alerts

* load JSON schemas inline for web support

* switch test library to vitest

* fix up imports and semver

try to reduce use of predicates when there are no child validators.
initial thoughts about a CsvValidationRule class.

* additional numeric checks

* revise based on walkthrough findings

Use addItemsWithLimit when adding errors or alerts to a validator's
list.
Move dynamic validators to CsvHelpers so that the CsvValidator file is
less cluttered.
Remove unused OriginalFlavorValidator type.
Add validation and tests for methodology to ensure use of enum values in
all cases.
Use namespaced import for CSV errors.
Use consistent implementation style for sets of validators based on
payer-plans.

* non-modifier rows may omit setting

Move setting validation from general row validators to non-modifier
validators.
Add tests for usage of optional callback functions.
Add test for using a File with JsonValidator.
Add test coverage reporting.

* change recommended node version to 20

* test cleanup

Add tests for blank lines, error limit, and data callback in CSV.

* CodePairMissingError applies to row

Constructor for this error no longer takes parameter for column, and
instead sets column to -1.

* Refactor objectFromKeysValues for improved performance

New implementation builds object in one loop. In local benchmark tests,
this roughly halved the execution time.

* fix up imports and semver

* csv: alert when file has no payer-specific charges

* partial rule implementation

starting in version 3.0.0, remove the estimated amount field. replace it
with median, 10th percentile, and 90th percentile.

* start implementing 2026 rule features

JSON schema and CSV validator for 3.0 has numerous changes:
- some new min lengths for strings
- drug unit field is now a number
- optional enum setting property for modifiers
- conditional that required estimate now requires the new median, 10th,
  and 90th percentile fields
- new count of allowed amount property can be zero or positive, and when
  zero, removes the conditional requirement for median, 10th, and 90th
  percentile
- that conditional now also applies even if a dollar amount is encoded
- notes are required when count is 0
- a payers information object must have some kind of charge
- CMG is new allowed code type
- a row in CSV tall with payer and plan name must have a payer-specific
  charge
- hospital_location is now called location_name
- affirmation is now attestation, and has an associated attester_name
- new type_2_npi general data element

* new attestation statement

Update attestation statement text in JSON schema and CSV validation
helper file.
Move some new elements around in the JSON schema to have fewer top-level
elements.

* add metadata callback for CSV

* new values for count of allowed amounts

The allowed values for count of allowed amounts are 0, integers 11 and
greater, and the phrase "1 through 10". In the JSON schema, the type is
string, so the numeric values should be quoted in the MRF.

Additional notes are required when count of allowed amounts is 0.

* add alert for false affirmation or attestation

* fix v3 schema

type 2 npi is a top-level element. collect it when collecting metadata
in the json validator.
fix alert imports for both validators.

fixing up type definition files